+++ Navigation.h
@@ -20,8 +20,8 @@ private:
     // Attribute für die Positionsbestimmung
     int X_Koordinaten[5] = {0, 1000, 0, 0, 25};
     int Y_Koordinaten[5] = {1000, 1000, 0, 0, 9};
-    int x_aktuell = 0;
-    int y_aktuell = 0;
+    float x_aktuell = 0;
+    float y_aktuell = 0;
     int maxPosition = 2;	// Maximal 5 Positionen
     int Position;				// Aktueller Anfahrpunkt
     // Reglungswerkzeuge
